Update temprature in app with API

+++ lib/app/domain/models/weather_model.dart
@@ -61,7 +61,7 @@ class Temperature with _$Temperature {
 @freezed
 class Minimum with _$Minimum {
   factory Minimum(
-    @JsonKey(name: "Value") int value,
+    @JsonKey(name: "Value") double value,
   ) = _Minimum;
 
   factory Minimum.fromJson(Map<String, dynamic> json) =>
